# contracts/openapi/ticket-order-api/openapi.yml
...
openapi: 3.1.0
info:
  title: Ticket Service API
  version: 0.1.0
  description: API contract for the TicketOrderPlatform ticket service.
servers:
- url: http://localhost:8080
  description: Local API server
...
paths:
...
  "/auth/register":
    post:
      operationId: registerUser
      tags:
      - Auth
      summary: Register a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RegisterUserRequest"
      responses:
        '201':
          description: Registration succeeded
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UserResponse"
        '400':
          description: Registration request was invalid
        '409':
          description: User email is already registered
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
...
    RegisterUserRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          minLength: 8
          maxLength: 128
...
    UserRole:
      type: string
      enum:
      - CUSTOMER
      - MANAGER
      - ADMIN
...
    UserResponse:
      type: object
      required:
      - id
      - email
      - role
      - enabled
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 320
        role:
          "$ref": "#/components/schemas/UserRole"
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time